ioss: move Iotr_Factory to its own header
In the right library of course.
