+++ packages/seacas/libraries/ioss/src/transform/Iotr_VectorMagnitude.h
@@ -10,6 +10,7 @@
 
 #include <Ioss_Transform.h>    // for Transform, Factory
 #include <Ioss_VariableType.h> // for VariableType
+#include <transform/Iotr_Factory.h>
 
 #include <string> // for string
